+++ src/main.js
@@ -8,8 +8,7 @@ import { pushData, openDataset } from './storages/dataset';
 import events, { initializeEvents, stopEvents } from './events';
 import { getValue, setValue, getInput, openKeyValueStore } from './storages/key_value_store';
 import { launchPuppeteer } from './puppeteer';
-import PuppeteerCrawler from './crawlers/puppeteer_crawler';
-import PuppeteerPool from './puppeteer_pool';
+import BrowserCrawler from './crawlers/browser_crawler';
 import Request from './request';
 import { RequestList, openRequestList } from './request_list';
 import { createProxyConfiguration } from './proxy_configuration';
@@ -78,8 +77,7 @@ export {
     openKeyValueStore,
 
     launchPuppeteer,
-    PuppeteerPool,
-    PuppeteerCrawler,
+    BrowserCrawler,
 
     PseudoUrl,
